#8583: fold handed into viewed and split the fork test out
