+++ eo-lowering/src/main/java/org/eolang/lowering/Rendering.java
@@ -281,27 +281,12 @@ public final class Rendering {
     }
 
     /**
-     * The Java expression the atom hands to {@code Data.ToPhi}.
-     *
-     * <p>Where the fragment settled into a view of a local rather than
-     * the local itself, as {@code x!} does, the step is a double and the
-     * answer is bytes: the raw bits of the local are those bytes.</p>
-     *
-     * @param local The Java expression of the value, such as {@code s1}
-     * @param key The key that value stands under, such as {@code sym:s1}
-     * @return The expression to hand over, wrapped when the formas part
-     */
-    public String handed(final String local, final String key) {
-        return this.viewed(local, key, Rendering.carried(this.program.carrier()));
-    }
-
-    /**
      * The Java expression of a value, seen as the forma another place expects.
      *
-     * <p>A local and the key it stands under may part, as {@code x!} makes
-     * them part: the local is a double and the key is bytes. Wherever such a
-     * value is handed over, returned or assigned, the bits have to be seen
-     * as the forma of the place that takes them.</p>
+     * <p>Where the fragment settled into a view of a local rather than the
+     * local itself, as {@code x!} does, the step is a double and the key is
+     * bytes. Wherever such a value is handed over, returned or assigned, the
+     * bits have to be seen as the forma of the place that takes them.</p>
      *
      * @param local The Java expression of the value, such as {@code s1}
      * @param key The key that value stands under, such as {@code sym:s1}
@@ -310,23 +295,24 @@ public final class Rendering {
      */
     public String viewed(final String local, final String key, final String carrier) {
         final String own = this.forma(key);
+        final String wanted = Rendering.carried(carrier);
         final String out;
-        if (carrier.equals(own)) {
+        if (wanted.equals(own)) {
             out = local;
-        } else if ("bytes".equals(carrier) && "number".equals(own)) {
+        } else if ("bytes".equals(wanted) && "number".equals(own)) {
             out = String.format(
                 "java.nio.ByteBuffer.allocate(8).putLong(Double.doubleToRawLongBits(%s)).array()",
                 local
             );
-        } else if ("bytes".equals(carrier) && "bool".equals(own)) {
+        } else if ("bytes".equals(wanted) && "bool".equals(own)) {
             out = String.format("new byte[] {(byte) (%s ? 0xFF : 0x00)}", local);
-        } else if ("number".equals(carrier) && "bytes".equals(own)) {
+        } else if ("number".equals(wanted) && "bytes".equals(own)) {
             out = String.format("java.nio.ByteBuffer.wrap(%s).getDouble()", local);
         } else {
             throw new IllegalStateException(
                 String.format(
                     "The answer '%s' carries a %s, which no view renders as a %s",
-                    key, own, carrier
+                    key, own, wanted
                 )
             );
         }
